chore(api): move InvalidAgentHarnessError to the workflow type module
Per the api layering rules, a workflow-domain exception belongs in
core/workflows/types.py beside the WorkflowError family, not in the service that
raises it. The FastAPI router and the platform tool handler both import it, so it
was crossing layers from the wrong home.

Pure move, no behavior change. The class keeps its exact inheritance, its code, and
its to_detail envelope; its private _json_safe_echo helper moves with it because
nothing else uses it and types.py must not import from service.py. RevisionConflictError
stays where it is: it predates this work.
